Simplifying InterferenceGraph abstraction
Since temporaries are all unique, we don't need the int binded to Node
on the graph abstracion implemented on TigerGraph. Having that int on
interference graphs nodes' only made the next stage more complicated.

I decided to change only InterferenceGraph and not all graphs to so that
the refactor was simpler.